Music: move global blockly setup to separate function
